# Convert functions needed for inference to pure tensorflow.