--- a/lib/layer_utils/proposal_top_layer.py
+++ b/lib/layer_utils/proposal_top_layer.py
@@ -9,8 +9,10 @@ from __future__ import print_function
 
 import numpy as np
 from model.config import cfg
-from model.bbox_transform import bbox_transform_inv, clip_boxes
+from model.bbox_transform import bbox_transform_inv, clip_boxes, bbox_transform_inv_tf, clip_boxes_tf
 import numpy.random as npr
+
+import tensorflow as tf
 
 def proposal_top_layer(rpn_cls_prob, rpn_bbox_pred, im_info, _feat_stride, anchors, num_anchors):
   """A layer that just selects the top region proposals
@@ -51,3 +53,21 @@ def proposal_top_layer(rpn_cls_prob, rpn_bbox_pred, im_info, _feat_stride, ancho
   batch_inds = np.zeros((proposals.shape[0], 1), dtype=np.float32)
   blob = np.hstack((batch_inds, proposals.astype(np.float32, copy=False)))
   return blob, scores
+
+def proposal_top_layer_tf(rpn_cls_prob, rpn_bbox_pred, im_info, _feat_stride, anchors, num_anchors):
+  rpn_top_n = cfg.TEST.RPN_TOP_N
+
+  scores = rpn_cls_prob[:, :, :, num_anchors:]
+  rpn_bbox_pred = tf.reshape(rpn_bbox_pred, shape=(-1, 4))
+  scores = tf.reshape(scores, shape=(-1,))
+
+  top_scores, top_inds = tf.nn.top_k(scores, k=rpn_top_n)
+  top_scores = tf.reshape(top_scores, shape=(-1, 1))
+  top_anchors = tf.gather(anchors, top_inds)
+  top_rpn_bbox = tf.gather(rpn_bbox_pred, top_inds)
+  proposals = bbox_transform_inv_tf(top_anchors, top_rpn_bbox)
+  proposals = clip_boxes_tf(proposals, im_info[:2])
+  proposals = tf.to_float(proposals)
+  batch_inds = tf.zeros((rpn_top_n, 1))
+  blob = tf.concat([batch_inds, proposals], 1)
+  return blob, top_scores