--- a/lib/layer_utils/snippets.py
+++ b/lib/layer_utils/snippets.py
@@ -7,6 +7,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import tensorflow as tf
 import numpy as np
 from layer_utils.generate_anchors import generate_anchors
 
@@ -27,3 +28,23 @@ def generate_anchors_pre(height, width, feat_stride, anchor_scales=(8,16,32), an
   length = np.int32(anchors.shape[0])
 
   return anchors, length
+
+def generate_anchors_tf(height, width, feat_stride=16, anchor_scales=(8, 16, 32), anchor_ratios=(0.5, 1, 2)):
+  shift_x = tf.range(width) * feat_stride # width
+  shift_y = tf.range(height) * feat_stride # height
+  shift_x, shift_y = tf.meshgrid(shift_x, shift_y)
+  sx = tf.reshape(shift_x, shape=(-1,))
+  sy = tf.reshape(shift_y, shape=(-1,))
+  shifts = tf.transpose(tf.stack([sx, sy, sx, sy]))
+  K = tf.multiply(width, height)
+  shifts = tf.transpose(tf.reshape(shifts, shape=[1, K, 4]), perm=(1, 0, 2))
+
+  anchors = generate_anchors(ratios=np.array(anchor_ratios), scales=np.array(anchor_scales))
+  A = anchors.shape[0]
+  anchor_constant = tf.constant(anchors.reshape((1, A, 4)), dtype=tf.int32)
+
+  length = K * A
+  anchors_tf = tf.reshape(tf.add(anchor_constant, shifts), shape=(length, 4))
+  
+  return tf.cast(anchors_tf, dtype=tf.float32), length
+