--- a/lib/layer_utils/snippets.py
+++ b/lib/layer_utils/snippets.py
@@ -11,25 +11,7 @@ import tensorflow as tf
 import numpy as np
 from layer_utils.generate_anchors import generate_anchors
 
-def generate_anchors_pre(height, width, feat_stride, anchor_scales=(8,16,32), anchor_ratios=(0.5,1,2)):
-  """ A wrapper function to generate anchors given different scales
-    Also return the number of anchors in variable 'length'
-  """
-  anchors = generate_anchors(ratios=np.array(anchor_ratios), scales=np.array(anchor_scales))
-  A = anchors.shape[0]
-  shift_x = np.arange(0, width) * feat_stride
-  shift_y = np.arange(0, height) * feat_stride
-  shift_x, shift_y = np.meshgrid(shift_x, shift_y)
-  shifts = np.vstack((shift_x.ravel(), shift_y.ravel(), shift_x.ravel(), shift_y.ravel())).transpose()
-  K = shifts.shape[0]
-  # width changes faster, so here it is H, W, C
-  anchors = anchors.reshape((1, A, 4)) + shifts.reshape((1, K, 4)).transpose((1, 0, 2))
-  anchors = anchors.reshape((K * A, 4)).astype(np.float32, copy=False)
-  length = np.int32(anchors.shape[0])
-
-  return anchors, length
-
-def generate_anchors_tf(height, width, feat_stride=16, anchor_scales=(8, 16, 32), anchor_ratios=(0.5, 1, 2)):
+def generate_anchors_pre(height, width, feat_stride=16, anchor_scales=(8, 16, 32), anchor_ratios=(0.5, 1, 2)):
   shift_x = tf.range(width) * feat_stride # width
   shift_y = tf.range(height) * feat_stride # height
   shift_x, shift_y = tf.meshgrid(shift_x, shift_y)