--- a/lib/model/bbox_transform.py
+++ b/lib/model/bbox_transform.py
@@ -9,6 +9,7 @@ from __future__ import division
 from __future__ import print_function
 
 import numpy as np
+import tensorflow as tf
 
 def bbox_transform(ex_rois, gt_rois):
   ex_widths = ex_rois[:, 2] - ex_rois[:, 0] + 1.0
@@ -78,3 +79,39 @@ def clip_boxes(boxes, im_shape):
   # y2 < im_shape[0]
   boxes[:, 3::4] = np.maximum(np.minimum(boxes[:, 3::4], im_shape[0] - 1), 0)
   return boxes
+
+
+
+def bbox_transform_inv_tf(boxes, deltas):
+  boxes = tf.cast(boxes, deltas.dtype)
+  widths = tf.subtract(boxes[:, 2], boxes[:, 0]) + 1.0
+  heights = tf.subtract(boxes[:, 3], boxes[:, 1]) + 1.0
+  ctr_x = tf.add(boxes[:, 0], widths * 0.5)
+  ctr_y = tf.add(boxes[:, 1], heights * 0.5)
+
+  dx = deltas[:, 0]
+  dy = deltas[:, 1]
+  dw = deltas[:, 2]
+  dh = deltas[:, 3]
+
+  pred_ctr_x = tf.add(tf.multiply(dx, widths), ctr_x)
+  pred_ctr_y = tf.add(tf.multiply(dy, widths), ctr_y)
+  pred_w = tf.multiply(tf.exp(dw), widths)
+  pred_h = tf.multiply(tf.exp(dh), heights)
+
+  pred_boxes0 = tf.subtract(pred_ctr_x, pred_w * 0.5)
+  pred_boxes1 = tf.subtract(pred_ctr_y, pred_h * 0.5)
+  pred_boxes2 = tf.add(pred_ctr_x, pred_w * 0.5)
+  pred_boxes3 = tf.add(pred_ctr_y, pred_h * 0.5)
+
+  return tf.stack([pred_boxes0, pred_boxes1, pred_boxes2, pred_boxes3], axis=1)
+
+
+def clip_boxes_tf(boxes, im_info):
+  b0 = tf.maximum(tf.minimum(boxes[:, 0], im_info[1] - 1), 0)
+  b1 = tf.maximum(tf.minimum(boxes[:, 1], im_info[0] - 1), 0)
+  b2 = tf.maximum(tf.minimum(boxes[:, 2], im_info[1] - 1), 0)
+  b3 = tf.maximum(tf.minimum(boxes[:, 3], im_info[0] - 1), 0)
+  return tf.stack([b0, b1, b2, b3], axis=1)
+
+