--- a/lib/nets/network.py
+++ b/lib/nets/network.py
@@ -14,7 +14,7 @@ from tensorflow.contrib.slim import arg_scope
 
 import numpy as np
 
-from layer_utils.snippets import generate_anchors_pre
+from layer_utils.snippets import generate_anchors_pre, generate_anchors_tf
 from layer_utils.proposal_layer import proposal_layer
 from layer_utils.proposal_top_layer import proposal_top_layer
 from layer_utils.anchor_target_layer import anchor_target_layer
@@ -189,10 +189,11 @@ class Network(object):
       # just to get the shape right
       height = tf.to_int32(tf.ceil(self._im_info[0] / np.float32(self._feat_stride[0])))
       width = tf.to_int32(tf.ceil(self._im_info[1] / np.float32(self._feat_stride[0])))
-      anchors, anchor_length = tf.py_func(generate_anchors_pre,
-                                          [height, width,
-                                           self._feat_stride, self._anchor_scales, self._anchor_ratios],
-                                          [tf.float32, tf.int32], name="generate_anchors")
+      anchors, anchor_length = generate_anchors_tf(height, width, self._feat_stride, self._anchor_scales, self._anchor_ratios)
+      # anchors, anchor_length = tf.py_func(generate_anchors_pre,
+      #                                     [height, width,
+      #                                      self._feat_stride, self._anchor_scales, self._anchor_ratios],
+      #                                     [tf.float32, tf.int32], name="generate_anchors")
       anchors.set_shape([None, 4])
       anchor_length.set_shape([])
       self._anchors = anchors