--- a/lib/nets/network.py
+++ b/lib/nets/network.py
@@ -14,9 +14,9 @@ from tensorflow.contrib.slim import arg_scope
 
 import numpy as np
 
-from layer_utils.snippets import generate_anchors_pre, generate_anchors_tf
-from layer_utils.proposal_layer import proposal_layer, proposal_layer_tf
-from layer_utils.proposal_top_layer import proposal_top_layer, proposal_top_layer_tf
+from layer_utils.snippets import generate_anchors_pre
+from layer_utils.proposal_layer import proposal_layer
+from layer_utils.proposal_top_layer import proposal_top_layer
 from layer_utils.anchor_target_layer import anchor_target_layer
 from layer_utils.proposal_target_layer import proposal_target_layer
 from utils.visualization import draw_bounding_boxes
@@ -87,7 +87,7 @@ class Network(object):
 
   def _proposal_top_layer(self, rpn_cls_prob, rpn_bbox_pred, name):
     with tf.variable_scope(name) as scope:
-      rois, rpn_scores = proposal_top_layer_tf(
+      rois, rpn_scores = proposal_top_layer(
         rpn_cls_prob,
         rpn_bbox_pred,
         self._im_info,
@@ -95,10 +95,6 @@ class Network(object):
         self._anchors,
         self._num_anchors
       )
-      # rois, rpn_scores = tf.py_func(proposal_top_layer,
-      #                               [rpn_cls_prob, rpn_bbox_pred, self._im_info,
-      #                                self._feat_stride, self._anchors, self._num_anchors],
-      #                               [tf.float32, tf.float32], name="proposal_top")
       rois.set_shape([cfg.TEST.RPN_TOP_N, 5])
       rpn_scores.set_shape([cfg.TEST.RPN_TOP_N, 1])
 
@@ -106,7 +102,7 @@ class Network(object):
 
   def _proposal_layer(self, rpn_cls_prob, rpn_bbox_pred, name):
     with tf.variable_scope(name) as scope:
-      rois, rpn_scores = proposal_layer_tf(
+      rois, rpn_scores = proposal_layer(
         rpn_cls_prob,
         rpn_bbox_pred,
         self._im_info,
@@ -115,10 +111,6 @@ class Network(object):
         self._anchors,
         self._num_anchors
       )
-      # rois, rpn_scores = tf.py_func(proposal_layer,
-      #                         [rpn_cls_prob, rpn_bbox_pred, self._im_info, self._mode,
-      #                          self._feat_stride, self._anchors, self._num_anchors],
-      #                         [tf.float32, tf.float32], name="proposal")
       rois.set_shape([None, 5])
       rpn_scores.set_shape([None, 1])
 
@@ -206,11 +198,13 @@ class Network(object):
       # just to get the shape right
       height = tf.to_int32(tf.ceil(self._im_info[0] / np.float32(self._feat_stride[0])))
       width = tf.to_int32(tf.ceil(self._im_info[1] / np.float32(self._feat_stride[0])))
-      anchors, anchor_length = generate_anchors_tf(height, width, self._feat_stride, self._anchor_scales, self._anchor_ratios)
-      # anchors, anchor_length = tf.py_func(generate_anchors_pre,
-      #                                     [height, width,
-      #                                      self._feat_stride, self._anchor_scales, self._anchor_ratios],
-      #                                     [tf.float32, tf.int32], name="generate_anchors")
+      anchors, anchor_length = generate_anchors_pre(
+        height,
+        width,
+        self._feat_stride,
+        self._anchor_scales,
+        self._anchor_ratios
+      )
       anchors.set_shape([None, 4])
       anchor_length.set_shape([])
       self._anchors = anchors