Currently, the code for `generate_anchors_pre`, `proposal_layer`, and `proposal_top_layer` use `tf.py_func`, which makes the model generated by this code unusable for export. This pull request implements these functions in pure tensorflow operations, so the model can be used on different devices and runtimes.

Using pure tensorflow also has a couple of other benefits, such as tensorflow's built in `optimize_for_inference` function. In my experience this cuts off 10-15% of the inference time.