@endernewton 2018-03-15
@Detry322 How does it affect the speed and performance? I am curious to learn more!