@Detry322 2018-03-16
In my initial CPU-only testing (don't have easy access to a GPU right now), `im_detect` takes about 5% more time to run using the tensorflow-only code.


Here are a couple runs to give you an idea on my CPU, generated with the following command, taking the taking the runtime after it has stabilized:

`./experiments/scripts/test_faster_rcnn.sh 0 pascal_voc_0712 res101`

Before patch:

```
im_detect: 10/4952 5.429s 0.005s
im_detect: 11/4952 5.440s 0.005s
im_detect: 12/4952 5.380s 0.005s
im_detect: 13/4952 5.417s 0.005s
im_detect: 14/4952 5.428s 0.005s
```

After patch:

```
im_detect: 10/4952 5.610s 0.007s
im_detect: 11/4952 5.631s 0.006s
im_detect: 12/4952 5.669s 0.006s
im_detect: 13/4952 5.691s 0.006s
im_detect: 14/4952 5.686s 0.006s
```

After patch with tensorflow's `optimize_for_inference`:

```
im_detect: 30/4952 4.515s 0.003s
im_detect: 31/4952 4.508s 0.003s
im_detect: 32/4952 4.505s 0.003s
im_detect: 33/4952 4.503s 0.003s
im_detect: 34/4952 4.500s 0.003s
```

I expect this code to be slower and more accurate than the original code. This is because the NMS code before had a two-stage process - cutting down proposals first by taking the first `pre_nms_topN` proposals, and then after nms taking the `post_nms_topN`. I believe this was done because you wanted to limit the number of proposals passed to `nms` to limit computation time.

The effect is, in the original code, sometimes fewer proposals than `post_nms_topN` would make it through both "top_N" filters, and so less time overall would be spent evaluating these proposals. In contrast, the tensorflow NMS implementation almost always selects exactly `post_nms_topN` proposals, so the runtime is more consistent. Since there's only a single filter, these proposals are also of higher quality, since good ones won't be dropped early. To make the code run faster but with similar accuracy, I suspect you could just lower `post_nms_topN`

Notice however, how after running "optimize_for_inference" on the graph the code is ~20% faster than the original!

I'll see if I can get access to a GPU to give you accuracy and speed numbers, but since I don't have one it'll be difficult. If you have access to one already, it might be easier for you to run it yourself.