@endernewton 2018-03-18
Great, thanks! Let me ping the people who had the need for pure tensorflow inferences. At the same time, could you give the AP numbers with the released model? With tensorflow nms does it afffect much of the detection results?