@Detry322 2018-03-19
I just got access to a GPU, so here's some speed numbers using `-arch=sm_37`.

## Pre-patch

Speed
```
im_detect: 4948/4952 0.361s 0.013s
im_detect: 4949/4952 0.361s 0.013s
im_detect: 4950/4952 0.361s 0.013s
im_detect: 4951/4952 0.361s 0.013s
im_detect: 4952/4952 0.361s 0.013s
```
Accuracy:
```
AP for aeroplane = 0.8303
AP for bicycle = 0.8598
AP for bird = 0.8149
AP for boat = 0.7416
AP for bottle = 0.6829
AP for bus = 0.8758
AP for car = 0.8798
AP for cat = 0.8835
AP for chair = 0.6238
AP for cow = 0.8682
AP for diningtable = 0.7029
AP for dog = 0.8840
AP for horse = 0.8726
AP for motorbike = 0.8292
AP for person = 0.8298
AP for pottedplant = 0.5289
AP for sheep = 0.8076
AP for sofa = 0.7752
AP for train = 0.8447
AP for tvmonitor = 0.7929
Mean AP = 0.7964
```

## Post-patch

Speed
```
im_detect: 4948/4952 0.366s 0.014s
im_detect: 4949/4952 0.366s 0.014s
im_detect: 4950/4952 0.366s 0.014s
im_detect: 4951/4952 0.366s 0.014s
im_detect: 4952/4952 0.366s 0.014s
```

Accuracy
```
AP for aeroplane = 0.8163
AP for bicycle = 0.8618
AP for bird = 0.8172
AP for boat = 0.7467
AP for bottle = 0.6867
AP for bus = 0.8736
AP for car = 0.8813
AP for cat = 0.8728
AP for chair = 0.6207
AP for cow = 0.8659
AP for diningtable = 0.7105
AP for dog = 0.8799
AP for horse = 0.8701
AP for motorbike = 0.8324
AP for person = 0.8302
AP for pottedplant = 0.5358
AP for sheep = 0.8198
AP for sofa = 0.7807
AP for train = 0.8588
AP for tvmonitor = 0.7927
Mean AP = 0.7977
```