@JRCondeNast 2018-03-20
@Detry322 Thanks for doing this. Just curious, did you test the im_detect performance inside demo.py and what the number looks like. I used your code and run the demo.py using the pascal_voc model(resnet/110000 iters) trained by Xin Lei and it took 17 sec for im_detect function to send back results. My computer spec: OS X EI Capitan 2.7GHz INTEL CORE i5 16GB 1867 MHz DDR3.